Move some IPC demo code to example folder under go-clipboard
