+++ go-clipboard/example/.gitignore
@@ -1,0 +1,2 @@
+./screenshot.png
+client
