+++ go-clipboard/example/cb-ops.js
